# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Localiza Integração Cenário 4
  description: API demonstrativa para CRM, reservas e contratos, financeiro e faturamento,
    atendimento e assistência 24h, workflow e integração.
  version: 1.0.0
paths:
...
  /api/v1/contracts/reservations:
    post:
      tags:
      - Reservas e contratos
      summary: Create Reservation
      description: Registra reserva para cliente elegível; com Idempotency-Key, repetir
        a chamada devolve a mesma reserva.
      operationId: create_reservation_api_v1_contracts_reservations_post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          description: Opcional; quando informado, repetir a chamada devolve a mesma
            reserva.
          title: Idempotency-Key
        description: Opcional; quando informado, repetir a chamada devolve a mesma
          reserva.
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
          headers:
            Idempotency-Replayed:
              description: Retorna true quando a API reutiliza a resposta armazenada.
              schema:
                type: string
                enum:
                - 'true'
                - 'false'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/contracts/reservations
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/contracts/reservations
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Cliente não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Cliente não encontrado
                instance: /api/v1/contracts/reservations
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: Idempotency-Key reutilizada com outra reserva.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:conflict
                title: Conflito
                status: 409
                detail: Idempotency-Key reutilizada com outra reserva
                instance: /api/v1/contracts/reservations
                code: CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Cliente inelegível, sem consentimento, inativo ou período inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail: Cliente inelegível, sem consentimento, inativo ou período
                  inválido
                instance: /api/v1/contracts/reservations
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
    get:
      tags:
      - Reservas e contratos
      summary: List Reservations
      description: Lista até 200 reservas, das mais recentes para as mais antigas,
        com filtro opcional por cliente.
      operationId: list_reservations_api_v1_contracts_reservations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: customerId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Customerid
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
                title: Response List Reservations Api V1 Contracts Reservations Get
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/contracts/reservations
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/contracts/reservations
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/contracts/reservations
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
  /api/v1/contracts/reservations/{reservation_id}:
    get:
      tags:
      - Reservas e contratos
      summary: Get Reservation
      description: Consulta uma reserva pelo identificador global.
      operationId: get_reservation_api_v1_contracts_reservations__reservation_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Reservation Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/contracts/reservations/{reservation_id}
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/contracts/reservations/{reservation_id}
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Reserva não encontrada.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Reserva não encontrada
                instance: /api/v1/contracts/reservations/{reservation_id}
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/contracts/reservations/{reservation_id}
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
  /api/v1/contracts/reservations/{reservation_id}/draft:
    post:
      tags:
      - Reservas e contratos
      summary: Draft From Reservation
      description: Gera o rascunho de contrato a partir da reserva, sem recadastrar
        o cliente, na mesma transação.
      operationId: draft_from_reservation_api_v1_contracts_reservations__reservation_id__draft_post
      security:
      - HTTPBearer: []
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Reservation Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          maxLength: 100
          title: Idempotency-Key
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDraftResponse'
          headers:
            Idempotency-Replayed:
              description: Retorna true quando a API reutiliza a resposta armazenada.
              schema:
                type: string
                enum:
                - 'true'
                - 'false'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/contracts/reservations/{reservation_id}/draft
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/contracts/reservations/{reservation_id}/draft
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Reserva não encontrada.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Reserva não encontrada
                instance: /api/v1/contracts/reservations/{reservation_id}/draft
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: Reserva cancelada, Idempotency-Key reutilizada ou contrato
            já existente para cliente, serviço e início.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:conflict
                title: Conflito
                status: 409
                detail: Reserva cancelada, Idempotency-Key reutilizada ou contrato
                  já existente para cliente, serviço e início
                instance: /api/v1/contracts/reservations/{reservation_id}/draft
                code: CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Cliente inelegível, sem consentimento ou entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail: Cliente inelegível, sem consentimento ou entrada inválida
                instance: /api/v1/contracts/reservations/{reservation_id}/draft
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
  /api/v1/contracts/reservations/{reservation_id}/cancel:
    post:
      tags:
      - Reservas e contratos
      summary: Cancel Reservation
      description: Cancela reserva solicitada ou com rascunho; o rascunho vinculado
        também é cancelado.
      operationId: cancel_reservation_api_v1_contracts_reservations__reservation_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Reservation Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/contracts/reservations/{reservation_id}/cancel
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/contracts/reservations/{reservation_id}/cancel
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Reserva não encontrada.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Reserva não encontrada
                instance: /api/v1/contracts/reservations/{reservation_id}/cancel
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: Reserva ativa ou encerrada não pode ser cancelada.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:conflict
                title: Conflito
                status: 409
                detail: Reserva ativa ou encerrada não pode ser cancelada
                instance: /api/v1/contracts/reservations/{reservation_id}/cancel
                code: CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/contracts/reservations/{reservation_id}/cancel
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
...
components:
  schemas:
...
    ReservationInput:
      properties:
        customerId:
          type: string
          format: uuid
          title: Customerid
        vehicleGroup:
          type: string
          maxLength: 50
          minLength: 1
          title: Vehiclegroup
        protectionCode:
          type: string
          maxLength: 50
          minLength: 1
          title: Protectioncode
        serviceCode:
          type: string
          maxLength: 80
          minLength: 1
          title: Servicecode
        startsOn:
          type: string
          format: date
          title: Startson
        endsOn:
          type: string
          format: date
          title: Endson
        amount:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d{0,2}0*$
          title: Amount
        currency:
          type: string
          pattern: ^[A-Z]{3}$
          title: Currency
          default: BRL
        billingCycle:
          type: string
          enum:
          - ONCE
          - MONTHLY
          - QUARTERLY
          - YEARLY
          title: Billingcycle
          default: MONTHLY
        slaHours:
          type: integer
          maximum: 720.0
          minimum: 1.0
          title: Slahours
          default: 8
      type: object
      required:
      - customerId
      - vehicleGroup
      - protectionCode
      - serviceCode
      - startsOn
      - endsOn
      - amount
      title: ReservationInput
      examples:
      - amount: 750.0
        billingCycle: ONCE
        currency: BRL
        customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
        endsOn: '2027-01-08'
        protectionCode: BASICA
        serviceCode: RENTAL-FLEX
        slaHours: 8
        startsOn: '2027-01-04'
        vehicleGroup: SUV-COMPACTO
    ReservationResponse:
      properties:
        reservationId:
          type: string
          format: uuid
          title: Reservationid
        customerId:
          type: string
          format: uuid
          title: Customerid
        vehicleGroup:
          type: string
          title: Vehiclegroup
        protectionCode:
          type: string
          title: Protectioncode
        serviceCode:
          type: string
          title: Servicecode
        startsOn:
          type: string
          format: date
          title: Startson
        endsOn:
          type: string
          format: date
          title: Endson
        amount:
          type: number
          title: Amount
        currency:
          type: string
          title: Currency
        billingCycle:
          type: string
          title: Billingcycle
        slaHours:
          type: integer
          title: Slahours
        status:
          type: string
          enum:
          - REQUESTED
          - DRAFTED
          - ACTIVE
          - CLOSED
          - CANCELLED
          title: Status
        contractId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Contractid
      type: object
      required:
      - reservationId
      - customerId
      - vehicleGroup
      - protectionCode
      - serviceCode
      - startsOn
      - endsOn
      - amount
      - currency
      - billingCycle
      - slaHours
      - status
      - contractId
      title: ReservationResponse
      examples:
      - amount: 750.0
        billingCycle: ONCE
        currency: BRL
        customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
        endsOn: '2027-01-08'
        protectionCode: BASICA
        reservationId: 15151515-1515-4151-8151-151515151515
        serviceCode: RENTAL-FLEX
        slaHours: 8
        startsOn: '2027-01-04'
        status: REQUESTED
        vehicleGroup: SUV-COMPACTO
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer